New analysis mode for finntreebank people: * no lowercase * no uppercases * no compounds * probably some ohter things
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@1467 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/proc-functionality.sh
@@ -63,4 +63,12 @@ fi
 if ! diff test.strings $srcdir/proc-caps-out4.strings ; then
     exit 1
 fi
+if ! ../../tools/src/hfst-proc/hfst-proc --cg --raw proc-caps.hfst.ol < $srcdir/proc-caps-in.strings > test.strings ; then
+    echo raw cg fail:
+    cat test.strings
+    exit 1
+fi
+if ! diff test.strings $srcdir/proc-caps-out5.strings ; then
+    exit 1
+fi
 rm test.strings
